feat: enhance scrolling and text rendering in detail page
- Updated `ratatui` dependency to include the "unstable" feature for advanced widget capabilities.
- Modified `Action::LineUp` and `Action::LineDown` to use `u16` instead of `u32` for scroll counts.
- Added `max_scroll` and `scroll` fields to `DetailPage` to manage vertical scrolling.
- Implemented smooth scrolling in the detail page using `Paragraph::scroll` and calculated the maximum scroll limit based on content height.
- Adjusted `update` method to handle `LineUp` and `LineDown` actions, ensuring scroll position stays within bounds.
- Reset scroll position and clear topic details in the `init` method when reinitializing the page.
- Improved rendering logic to dynamically calculate line count and constrain scrolling to visible content.

+++ src/action.rs
@@ -9,8 +9,8 @@ pub enum Action {
     FetchTopics,
     FetchTopicDetail(String),
     GoHome,
-    LineUp(u32),
-    LineDown(u32),
+    LineUp(u16),
+    LineDown(u16),
     OpenBrowser(String),
     Quit,
 }
